Provide binary release through GitHub actions (#45)
## Description

This PR targets users of ADORe by providing a binary release via GitHub
actions.

In a first version this release includes the latest successful build per
branch per platform. The action creates a tag called latest and attached
the releases as an archive with a static link like
https://github.com/<owner>/adore/releases/download/latest/adore_<clean_branch_name>_latest_build_<platform>.txz.
This allows users to always pull the latest version.

The archive contains the ros2_workspace/install folder, the setup.sh
script for sourcing, the adore_scenarios including simulation_scenarios
and tools/adore_api. This allows a user to create a minimal environment
as
[Dockerfile](https://github.com/user-attachments/files/21698369/Dockerfile.txt)
to run ADORe.

This change restructures the actions, i.a. by using templates to avoid
repetition for different platforms. This shows minor issues:

- There is one warning for using jlumbroso/free-disk-space@main with the
arm64 image, but it is non-critical. That external action seems to be
unmaintained and not officially adapted to arm64.
- Sometimes the x86 CI test fails for an unknown reason. The arm64 CI
test always succeeds.
- The ADORe API currently doesn't work because it's not using relative
paths to the setup.sh script.

For the future I'm working on a ready to use image and caching during
the build process.

Documentation is still required, technical as well as a fitting
README.md to be supplied with the release. The current README.md is made
for developers and building ADORe from scratch.

## Type of Change

Please delete options that are not relevant:

- [x] CI

## Checklist

- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] Tests pass locally with my changes.
- [ ] I have updated the documentation where necessary.

## How Has This Been Tested?

- https://github.com/bioxz/adore/actions/runs/16849462258
- https://github.com/bioxz/adore/releases/tag/latest