adore_api port updated and bak file removed (#64)
## Description

Please provide a brief summary of the changes you made and why they are
necessary.

For the API of ADORe the port 5000 referenced in the documentation was
incorrect. It has been updated to 8888, the correct one.
An unneccessary bak file in the technical reference manual was deleted.

## Type of Change

Please delete options that are not relevant:

- [x] Documentation update (adds or updates documentation)

## Checklist

- [x] I have performed a self-review of my own code.
- [x] I have signed and submitted the Eclipse Foundation Contributors
Agreement: https://www.eclipse.org/legal/eca/.
- [x] My last commit was made with the `--signoff` flag as required by
the Eclipse Foundation.
- [ ] I have commented my code, particularly in hard-to-understand areas
and provided a README.md when necessary.
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [ ] Tests pass locally with my changes: (run 'make test' on the ADORe
project root to run unit tests).
- [x] I have updated the documentation in
`documentation/technical_reference_manual` where necessary.
- [ ] I have updated the `THIRD-PARTY.md` if I have introduced new
third-party libraries or dependencies to the project.
- [x] I have updated the `CONTRIBUTERS.md` if I wish to acknowledged for
my contribution.

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide
instructions so others can reproduce.

- Markdown files have been verified after push to forked repo and on
local machine.
- Repo was checked if .bak file deleted correctly.

**Test Configuration**:
- OS: Fedora 42
- Docker Version: Docker version 28.5.1, build e180ab8

---------

Signed-off-by: Niklas Böger <jan.boeger@dlr.de>