The foxglove bridge as of version `3.2.1-1noble` uploaded on 08.01.2025 will not communicate with Lichtblick-Suite.
Lichtblick in the browser reports the following error:
```
Check that the WebSocket server at ws://localhost:8765 is reachable and supports protocol version foxglove.websocket.v1.
```

netstat -plunt | grep 8765 reports that there is a service listening on
8765:
```
tcp        0      0 0.0.0.0:8765            0.0.0.0:*               LISTEN      2287784/foxglove_b
```
Despite this lichtblick refused to use the service reporting the
following error:
```
HTTP/1.1 400 Bad Request
Missing expected sec-websocket-protocol header
```

In order to address this the `adore_scenarios/scenario_helpers/visualizer.py`
was modified to use the rosbridge and the rosapi instead of the
foxglove_bridge. All documentaiotn was updated to refelct the change in
port from 8765 or the foxglove_bridge to 9090 for the rosbridge.

- Removed DLR copywrited images for eclipse foundation ip compliance
- Fixed minor syntax error in `.test` with extra "}" at the end of the
  file

Signed-off-by: Andrew Koerner <andrew.koerner@dlr.de>